+++ FinanceApp.ServerTests/RequestTest.cs
@@ -13,7 +13,7 @@ public class RequestTest
 {
     private static readonly CreateTransaction TestRequest = new()
     {
-        Value = new RequestField<int>
+        Value = new RequestField<long>
         {
             Value = 100
         },
